Remove VariantCall.end
I looked more closely at VCF and determined that "end" is not required and the way we were using
was ambiguous. As it turns out, the `userId` plus `chromosome:version:start:sampleType:sampleId`
is sufficient to uniquely identify a variantCall.

The reason for this is that VCF represents actual a high level concept (hence the name variant CALL),
not raw potentially overlapping sequence data. So the software that produces the VCF has resolved
overlaps and presents a single call at a position.